Split VariantCaster from binder_common.h
