+++ core/crypto/hashing_context.h
@@ -31,7 +31,7 @@
 #pragma once
 
 #include "core/object/ref_counted.h"
-#include "core/variant/binder_common.h"
+#include "core/variant/type_info.h"
 
 class HashingContext : public RefCounted {
 	GDCLASS(HashingContext, RefCounted);
