+++ core/debugger/local_debugger.cpp
@@ -34,6 +34,8 @@
 #include "core/os/main_loop.h"
 #include "core/os/os.h"
 
+#include <cstdio>
+
 struct LocalDebugger::ScriptsProfiler {
 	struct ProfileInfoSort {
 		bool operator()(const ScriptLanguage::ProfilingInfo &A, const ScriptLanguage::ProfilingInfo &B) const {
