+++ core/io/file_access.h
@@ -36,7 +36,7 @@
 #include "core/os/memory.h"
 #include "core/string/ustring.h"
 #include "core/typedefs.h"
-#include "core/variant/binder_common.h"
+#include "core/variant/type_info.h"
 
 /**
  * Multi-Platform abstraction for accessing to files.
