+++ core/io/xml_parser.h
@@ -33,7 +33,7 @@
 #include "core/object/ref_counted.h"
 #include "core/string/ustring.h"
 #include "core/templates/vector.h"
-#include "core/variant/binder_common.h"
+#include "core/variant/type_info.h"
 
 /*
   Based on irrXML (see their zlib license). Added mainly for compatibility with their Collada loader.
