+++ core/object/make_virtuals.py
@@ -197,7 +197,9 @@ def run(target, source, env):
 #pragma once
 
 #include "core/object/script_instance.h"
-#include "core/variant/binder_common.h"
+#include "core/variant/method_ptrcall.h"
+#include "core/variant/variant_caster.h"
+#include "core/variant/variant_internal.h"
 
 inline constexpr uintptr_t _INVALID_GDVIRTUAL_FUNC_ADDR = static_cast<uintptr_t>(-1);
