+++ editor/export/editor_export_preset.h
@@ -33,7 +33,7 @@
 class EditorExportPlatform;
 
 #include "core/object/ref_counted.h"
-#include "core/variant/binder_common.h"
+#include "core/variant/type_info.h"
 
 class EditorExportPreset : public RefCounted {
 	GDCLASS(EditorExportPreset, RefCounted);
