+++ modules/navigation_2d/nav_link_2d.h
@@ -32,7 +32,6 @@
 
 #include "2d/nav_base_iteration_2d.h"
 #include "core/os/rw_lock.h"
-#include "core/variant/binder_common.h"
 #include "nav_base_2d.h"
 #include "nav_utils_2d.h"
