+++ platform/linuxbsd/crash_handler_linuxbsd.cpp
@@ -49,6 +49,7 @@
 #include <execinfo.h>
 #include <link.h>
 #include <csignal>
+#include <cstdio>
 #include <cstdlib>
 
 static void handle_crash(int sig) {
