+++ platform/macos/crash_handler_macos.mm
@@ -39,6 +39,7 @@
 #include "main/main.h"
 
 #include <unistd.h>
+#include <cstdio>
 
 #if defined(DEBUG_ENABLED)
 #define CRASH_HANDLER_ENABLED 1
