+++ servers/camera/camera_server.h
@@ -34,7 +34,7 @@
 #include "core/object/ref_counted.h"
 #include "core/os/thread_safe.h"
 #include "core/templates/rid.h"
-#include "core/variant/binder_common.h"
+#include "core/variant/type_info.h"
 
 /**
 	The camera server is a singleton object that gives access to the various
