+++ servers/display/accessibility_server.h
@@ -31,7 +31,7 @@
 #pragma once
 
 #include "core/object/object.h"
-#include "core/variant/binder_common.h"
+#include "core/variant/type_info.h"
 #include "servers/display/accessibility_server_enums.h"
 #include "servers/display/display_server_enums.h"
