+++ servers/display/native_menu.h
@@ -32,8 +32,8 @@
 
 #include "core/object/object.h"
 #include "core/os/keyboard.h"
-#include "core/variant/binder_common.h"
 #include "core/variant/callable.h"
+#include "core/variant/type_info.h"
 
 class Texture2D;
