+++ servers/navigation_2d/navigation_path_query_result_2d.h
@@ -31,7 +31,7 @@
 #pragma once
 
 #include "core/object/ref_counted.h"
-#include "core/variant/binder_common.h"
+#include "core/variant/type_info.h"
 #include "core/variant/typed_array.h"
 #include "servers/navigation_2d/navigation_constants_2d.h"
