+++ servers/rendering/rendering_device.h
@@ -36,7 +36,7 @@
 #include "core/templates/local_vector.h"
 #include "core/templates/rb_map.h"
 #include "core/templates/rid_owner.h"
-#include "core/variant/binder_common.h"
+#include "core/variant/type_info.h"
 #include "core/variant/typed_array.h"
 #include "servers/display/display_server_enums.h"
 #include "servers/rendering/rendering_device_commons.h"
