+++ servers/xr/xr_pose.h
@@ -31,7 +31,7 @@
 #pragma once
 
 #include "core/object/ref_counted.h"
-#include "core/variant/binder_common.h"
+#include "core/variant/type_info.h"
 
 class XRPose : public RefCounted {
 	GDCLASS(XRPose, RefCounted);
