+++ servers/xr/xr_server.h
@@ -32,7 +32,7 @@
 
 #include "core/object/ref_counted.h"
 #include "core/os/thread_safe.h"
-#include "core/variant/binder_common.h"
+#include "core/variant/type_info.h"
 #include "core/variant/variant.h"
 
 class XRInterface;
